ajustes crud

+++ Microservicio_WasteManager/src/main/java/com/example/microserviciowastemanager/repositorios/WasteManagerRepository.java
@@ -6,6 +6,6 @@ import com.example.microserviciowastemanager.entidades.*;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface WasteManagerRepository extends JpaRepository<WasteManagerEntity, Integer> {
+public interface WasteManagerRepository extends JpaRepository<WasteManagerEntity, Long> {
 
 }
