+++ Microservicio_WasteManagerAddress/src/main/java/com/example/microserviciowastemanageraddres/repositorio/WasteManagerAddressRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.microserviciowastemanageraddres.entidades.*;
 
-public interface WasteManagerAddressRepository  extends JpaRepository<WasteManagerAddressEntity, Integer>{
+public interface WasteManagerAddressRepository  extends JpaRepository<WasteManagerAddressEntity, Long>{
 
 }
